+++ core/dialog/dialog_test.go
@@ -482,6 +482,7 @@ func (c *recordingCanvas) DrawText(text string, bounds geometry.Rect, fontSize f
 
 func (c *recordingCanvas) DrawImage(_ image.Image, _ geometry.Point) {}
 func (c *recordingCanvas) PushClip(_ geometry.Rect)                  {}
+func (c *recordingCanvas) PushClipRoundRect(_ geometry.Rect, _ float32) {}
 func (c *recordingCanvas) PopClip()                                  {}
 func (c *recordingCanvas) PushTransform(_ geometry.Point)            {}
 func (c *recordingCanvas) PopTransform()                             {}